Pull cache clearing functionality into its own module, out of the models.
The idea here is that caching could be used in other places, like the resources, with great effect, but it would be best if the caching machinery were separated from model.save.  Thinking specifically about submission sets, which are currently being cached in memory on each resource.  That cache should be shared across resources, and should be in the caching backend, instead of in memory.